+++ client/app/Navigation/UserMenu/UserSettings/context.ts
@@ -1,5 +1,9 @@
 import React from 'react'
 
 export const UserSettingsContext = React.createContext<{
-  onUpdateUserSettings: (key: string, value: string | boolean) => void
+  onUpdateUserSettings: (
+    key: string,
+    value: string | boolean,
+    reloadAfterSave?: boolean
+  ) => void
 }>(null)