+++ client/app/Navigation/UserMenu/UserSettings/types.tsx
@@ -1,6 +1,19 @@
-import { IUserSetting } from './USER_SETTINGS'
+import { IDropdownOption } from 'office-ui-fabric-react'
 
 export interface IUserSettingInputProps {
   user: any
   setting: IUserSetting
+}
+
+export interface IUserSetting {
+  key: string
+  label: string
+  type: 'dropdown' | 'bool'
+  description?: string
+  defaultValue?: any
+  reloadAfterSave?: boolean
+}
+
+export interface IUserSettingDropdown extends IUserSetting {
+  options: IDropdownOption[]
 }