+++ client/hooks/user/usePermissions.ts
@@ -1,7 +1,8 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable tsdoc/syntax */
 import { AppContext } from 'AppContext'
 import getPermissions, { PERMISSION } from 'config/security/permissions'
-import { useContext } from 'react'
+import { useContext, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { contains } from 'underscore'
 
@@ -26,9 +27,13 @@ export function usePermissions(permissionIds?: string[]) {
     permissions = permissions.filter((perm) => contains(permissionIds, perm.id))
   }
 
-  return {
-    permissions,
-    hasPermission: (permission: PERMISSION) =>
-      context?.user ? context.user.hasPermission(permission) : false
-  }
+  return useMemo(
+    () => ({
+      permissions,
+      hasPermission: (permission: PERMISSION) => {
+        return context?.user ? context.user.hasPermission(permission) : false
+      }
+    }),
+    [context?.user]
+  )
 }