+++ client/pages/usePages.tsx
@@ -10,7 +10,7 @@ import {
   ReportsPage,
   TimesheetPage
 } from 'pages'
-import React, { useMemo } from 'react'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { pick } from 'underscore'
 import { IPageComponent } from './types'
@@ -20,61 +20,59 @@ type UsePages = {
   nav: INavItemProps[]
 }
 
-export function usePages() {
+export function usePages(): UsePages {
   const { t } = useTranslation()
   const { hasPermission } = usePermissions()
-  return useMemo<UsePages>(() => {
-    const pages: IPageComponent[] = [
-      {
-        text: t('navigation.timesheet'),
-        iconName: 'TimeSheet',
-        path: '/timesheet',
-        hidden: !hasPermission(PERMISSION.ACCESS_TIMESHEET),
-        component: <TimesheetPage />
-      },
-      {
-        text: t('navigation.customers'),
-        iconName: 'People',
-        path: '/customers',
-        hidden: !hasPermission(PERMISSION.ACCESS_CUSTOMERS),
-        component: <CustomersPage />
-      },
-      {
-        text: t('navigation.projects'),
-        iconName: 'ProjectCollection',
-        path: '/projects',
-        hidden: !hasPermission(PERMISSION.ACCESS_PROJECTS),
-        component: <ProjectsPage />
-      },
-      {
-        text: t('navigation.reports'),
-        iconName: 'ReportDocument',
-        path: '/reports',
-        hidden: !hasPermission(PERMISSION.ACCESS_REPORTS),
-        component: <ReportsPage />
-      },
-      {
-        text: t('navigation.admin'),
-        iconName: 'Settings',
-        path: '/admin',
-        hidden: !hasPermission(PERMISSION.ACCESS_ADMIN),
-        component: <AdminPage />
-      },
-      {
-        path: '/',
-        component: <Home />
-      }
-    ]
-    return {
-      pages,
-      nav: pages
-        .filter((page) => !!page.text)
-        .map((page) => ({
-          text: undefined,
-          iconName: undefined,
-          ...pick(page, 'text', 'iconName', 'hidden', 'sections'),
-          to: page.path
-        }))
+  const pages: IPageComponent[] = [
+    {
+      text: t('navigation.timesheet'),
+      iconName: 'TimeSheet',
+      path: '/timesheet',
+      hidden: !hasPermission(PERMISSION.ACCESS_TIMESHEET),
+      component: <TimesheetPage />
+    },
+    {
+      text: t('navigation.customers'),
+      iconName: 'People',
+      path: '/customers',
+      hidden: !hasPermission(PERMISSION.ACCESS_CUSTOMERS),
+      component: <CustomersPage />
+    },
+    {
+      text: t('navigation.projects'),
+      iconName: 'ProjectCollection',
+      path: '/projects',
+      hidden: !hasPermission(PERMISSION.ACCESS_PROJECTS),
+      component: <ProjectsPage />
+    },
+    {
+      text: t('navigation.reports'),
+      iconName: 'ReportDocument',
+      path: '/reports',
+      hidden: !hasPermission(PERMISSION.ACCESS_REPORTS),
+      component: <ReportsPage />
+    },
+    {
+      text: t('navigation.admin'),
+      iconName: 'Settings',
+      path: '/admin',
+      hidden: !hasPermission(PERMISSION.ACCESS_ADMIN),
+      component: <AdminPage />
+    },
+    {
+      path: '/',
+      component: <Home />
     }
-  }, [])
+  ]
+  return {
+    pages,
+    nav: pages
+      .filter((page) => !!page.text)
+      .map((page) => ({
+        text: undefined,
+        iconName: undefined,
+        ...pick(page, 'text', 'iconName', 'hidden', 'sections'),
+        to: page.path
+      }))
+  }
 }