+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for initial signin with email #884
 - Showing more details in the summary view on hover #886
 - Summary view moved to Reports
+- Support for start page user setting
 
 ### Fixed
 - Issue where sunday hours would not be displayed in the UI after confirming a week #785